examples/simulator/requirements.txt
azure-ai-evaluation
azure-ai-projects
semantic-kernel
wikipedia